ecommerce/requirements.txt
appdirs
asgiref
black
certifi
cffi
chardet
click
coverage
cryptography
Django
django-debug-toolbar
django-js-asset
django-mptt
idna
isort
mypy-extensions
pathspec
paypal-checkout-serversdk
paypalhttp
Pillow
pycparser
pyOpenSSL
pytz
regex
requests
six
sqlparse
toml
typed-ast
typing-extensions
urllib3
Serializers
djangorestframework


